+++ src/app/register/update/page.jsx
@@ -1,0 +1,17 @@
+"use client";
+
+import RecordForm from "@/components/RecordForm";
+import { useSearchParams } from "next/navigation";
+
+export default function ActualizarRegistro(){
+
+    const searchParams = useSearchParams();
+    const idParam = searchParams.get("id")
+
+    return (
+        <div>
+            <h1 className="text-2xl font-bold mb-6">Actualizar Registro de Marca</h1>
+            <RecordForm id={idParam}></RecordForm>
+        </div>
+    )
+}
